Copy-on-Write Trunk
This changeset implements copy-on-write for trunk nodes, which includes
several high-level changes. This PR still needs to be rebased onto main,
but the the purpose is to discuss high- and low-level design decisions.

Changes in this PR:

Trunk root lock. A distributed RW lock is used to access/change the
current root.

Flush from root. Flushes proceed from the root and cascade immediately
rather than being triggered at the beginning of trunk_compact_bundle.

Copy-on-write. Trunk nodes cannot be modified directly, and instead are
change via a copy-on-write of the root-to-node path together with a
change of the root node.

Garbage Collection for unlinked branches and filters. After a
copy-on-write, the nodes on the old path will be unreferenced. This PR
does not GC the trunk nodes themselves, but it includes a GC path to
dereference the replaced branches and filters.

platform_batch_rwlock. Replaces distributed locks using dummy cache
pages with a batched distributed RW lock implementation in
platform.[ch].