(#500) Move hook-related global vars to task_system{} struct.
This commit removes the dependency of task system structures on global
variables declared in task.c . The hook-related variables are now
moved as members of the task_system{} struct. This removes accessing
potentially stale values when task-system is destroyed and re-created.
Also, TASK_MAX_HOOKS is now decreased from 8 to 4. This change largely
has no functional impact, and is mostly a test-stabilization fix.