Identify Memtable v/s Branch page types via BTree-print routines.
This commit extends BTree-print routines to also report the page
type, whether it's a branch or a memtable BTree. As the structures and
print methods are shared between two objects, this extra information
will help in diagnostics. Trunk nodes are likewise identified.
Minor fix in trunk_print_pivots() to align outputs for pivot key's string.